requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai>=2.9.0
python-dotenv==1.0.0
python-dateutil==2.8.2
jinja2==3.1.2
aiofiles==23.2.1

